find_plink {pbdRPC}	R Documentation

Find plink

Description

Find the plink internal compiled with pbdRPC package.

Usage


find_plink()

Details

The function returns a full path to the plink or plink.exe command if found.

Value

A full path a full path to the plink or plink.exe command if found. Otherwise, "plink" is returned.

Examples


## Not run: 
library(pbdRPC, quietly = TRUE)

find_plink()

## End(Not run)


[Package pbdRPC version 0.1-1 Index]