rpc_options {pbdRPC}	R Documentation

Set and Get Options for Remote Procedure Call

Description

Set and get default options of login information for rpc().

Usage


rpcopt_set(user = NULL, hostname = NULL, pport = NULL, exec.type = NULL,
  args = NULL, priv.key = NULL, priv.key.ppk = NULL)

rpcopt_get()

Arguments

user
user id for logging to the server.

hostname
the server ip or host name.

pport
ssh port opened on the server.

exec.type
either "ssh" or "plink".

args
arguments to the ssh or plink.exe.

priv.key
file name/path to the private key in OpenSSH format.

priv.key.ppk
file name/path to the private key in PuTTY format.

Details

rpcopt_set() is to alter default options of login information. The user defined options are set to .pbd_env$RPC.LI. rpcopt_get() is to get default options of login information from .pbd_env$RPC.LI.

Value

A list is returned.

See Also

rpc().

Examples


## Not run: 
library(pbdRPC, quietly = TRUE)
rpcopt_set(user = "snoweye", hostname = "192.168.56.101")

rpcopt_get()

## End(Not run)


[Package pbdRPC version 0.1-1 Index]